Move parser code from Core to Cli (CLI-exclusive logic)
The highlight parser is used only by the CLI and does not need to be
shared with the server. Moving it to SunnySunday.Cli is the correct
separation per the client/server architecture.

- Move src/SunnySunday.Core/Parsing/ → src/SunnySunday.Cli/Parsing/
- Update namespaces: SunnySunday.Core.Parsing → SunnySunday.Cli.Parsing
- Move Microsoft.Extensions.Logging.Abstractions dep: Core → Cli
- Update test using directive accordingly
- Update docs (quickstart.md, plan.md, data-model.md, tasks.md) to
  reflect the Cli project as the parser home
