+++ specs/003-highlight-parser/plan.md
@@ -5,7 +5,7 @@
 
 ## Summary
 
-Parse Kindle `My Clippings.txt` files into structured highlight data. The parser is a pure function in `SunnySunday.Core/Parsing/` that reads the file line-by-line, extracts book/author/highlight/metadata, deduplicates by exact `(title, author, text)` match, groups highlights by book, and returns an immutable `ParseResult`. Notes are treated as highlights with a `[my note]` prefix. Malformed entries are logged as warnings. No database, no HTTP, no side effects.
+Parse Kindle `My Clippings.txt` files into structured highlight data. The parser is a pure function in `SunnySunday.Cli/Parsing/` that reads the file line-by-line, extracts book/author/highlight/metadata, deduplicates by exact `(title, author, text)` match, groups highlights by book, and returns an immutable `ParseResult`. Notes are treated as highlights with a `[my note]` prefix. Malformed entries are logged as warnings. No database, no HTTP, no side effects.
 
 ## Technical Context
 
@@ -14,7 +14,7 @@ Parse Kindle `My Clippings.txt` files into structured highlight data. The parser
 **Storage**: N/A — parser is a pure function with no persistence
 **Testing**: xUnit (via `SunnySunday.Tests`)
 **Target Platform**: Cross-platform (.NET 10 runtime)
-**Project Type**: Library (shared logic in `SunnySunday.Core`)
+**Project Type**: CLI-exclusive logic in `SunnySunday.Cli`
 **Performance Goals**: 10,000 clippings parsed, deduplicated, and grouped in < 5 seconds
 **Constraints**: Single-pass streaming, no external dependencies beyond .NET BCL
 **Scale/Scope**: Typical file: 500–5,000 clippings; stress test: 50,000+ clippings
@@ -25,7 +25,7 @@ Parse Kindle `My Clippings.txt` files into structured highlight data. The parser
 
 | Principle | Status | Notes |
 |-----------|--------|-------|
-| I. Client/Server Separation | **PASS** | Parser lives in `Core` (shared); used by CLI, not server |
+| I. Client/Server Separation | **PASS** | Parser lives in `Cli` (CLI-exclusive); not shared with server |
 | II. CLI-First, No GUI | **PASS** | No UI in this feature; parser is logic-only |
 | III. Zero-Config Onboarding | **PASS** | No configuration required; parser takes a file path |
 | IV. Local Processing Only | **PASS** | Pure local file parsing, no network calls |
@@ -53,26 +53,29 @@ specs/003-highlight-parser/
 ### Source Code (repository root)
 
 ```text
-src/SunnySunday.Core/
-├── Models/                    # Existing persistence models (UNCHANGED)
-│   ├── Author.cs
-│   ├── Book.cs
-│   ├── Highlight.cs
-│   ├── Settings.cs
-│   └── User.cs
-└── Parsing/                   # NEW — parser types and logic
+src/SunnySunday.Cli/
+├── Program.cs
+└── Parsing/                   # Parser types and logic
     ├── RawClipping.cs         # Internal intermediate parse result
     ├── ParsedHighlight.cs     # Public: single highlight in output
     ├── ParsedBook.cs          # Public: book with grouped highlights
     ├── ParseResult.cs         # Public: top-level result container
     └── ClippingsParser.cs     # Public: static parser entry point
 
+src/SunnySunday.Core/
+└── Models/                    # Existing persistence models (UNCHANGED)
+    ├── Author.cs
+    ├── Book.cs
+    ├── Highlight.cs
+    ├── Settings.cs
+    └── User.cs
+
 src/SunnySunday.Tests/
 └── Parsing/                   # NEW — parser unit tests
     └── ClippingsParserTests.cs
 ```
 
-**Structure Decision**: Parser code goes into a new `Parsing/` namespace within the existing `SunnySunday.Core` project. No new .csproj files. Tests go into the existing `SunnySunday.Tests` project under a `Parsing/` folder. This matches the existing project structure and avoids adding a 5th project (YAGNI).
+**Structure Decision**: Parser code goes into a new `Parsing/` namespace within the existing `SunnySunday.Cli` project (CLI-exclusive, not shared with server). Tests go into the existing `SunnySunday.Tests` project under a `Parsing/` folder. This avoids adding a 5th project (YAGNI).
 
 ## Complexity Tracking
