+++ src/SunnySunday.Cli/Parsing/ClippingsParser.cs
@@ -2,7 +2,7 @@ using System.Globalization;
 using System.Text.RegularExpressions;
 using Microsoft.Extensions.Logging;
 
-namespace SunnySunday.Core.Parsing;
+namespace SunnySunday.Cli.Parsing;
 
 /// <summary>
 /// Parses Kindle "My Clippings.txt" files into structured highlight data.
