+++ src/SunnySunday.Cli/Parsing/ParseResult.cs
@@ -1,0 +1,9 @@
+﻿namespace SunnySunday.Cli.Parsing;
+
+/// <summary>
+/// The complete output of parsing a Kindle clippings file.
+/// </summary>
+public record ParseResult(
+    IReadOnlyList<ParsedBook> Books,
+    int TotalEntriesProcessed,
+    int DuplicatesRemoved);
