+++ src/SunnySunday.Cli/Parsing/ParsedHighlight.cs
@@ -1,0 +1,10 @@
+﻿namespace SunnySunday.Cli.Parsing;
+
+/// <summary>
+/// A single parsed highlight or note from a Kindle clippings file.
+/// Notes have their text prefixed with "[my note] ".
+/// </summary>
+public record ParsedHighlight(
+    string Text,
+    string? Location,
+    DateTimeOffset? AddedOn);
