+++ src/SunnySunday.Cli/Parsing/RawClipping.cs
@@ -1,0 +1,9 @@
+﻿namespace SunnySunday.Cli.Parsing;
+
+internal record RawClipping(
+    string Title,
+    string? Author,
+    bool IsNote,
+    string? Location,
+    DateTimeOffset? AddedOn,
+    string Text);
