+++ src/SunnySunday.Cli/SunnySunday.Cli.csproj
@@ -5,6 +5,7 @@
   </ItemGroup>
 
   <ItemGroup>
+    <PackageReference Include="Microsoft.Extensions.Logging.Abstractions" Version="10.0.6" />
     <PackageReference Include="Spectre.Console" Version="0.55.0" />
   </ItemGroup>
