--- src/SunnySunday.Core/Parsing/ParseResult.cs
@@ -1,9 +1,0 @@
-namespace SunnySunday.Core.Parsing;
-
-/// <summary>
-/// The complete output of parsing a Kindle clippings file.
-/// </summary>
-public record ParseResult(
-    IReadOnlyList<ParsedBook> Books,
-    int TotalEntriesProcessed,
-    int DuplicatesRemoved);
