--- src/SunnySunday.Core/Parsing/RawClipping.cs
@@ -1,9 +1,0 @@
-namespace SunnySunday.Core.Parsing;
-
-internal record RawClipping(
-    string Title,
-    string? Author,
-    bool IsNote,
-    string? Location,
-    DateTimeOffset? AddedOn,
-    string Text);
