+++ src/SunnySunday.Core/SunnySunday.Core.csproj
@@ -7,8 +7,4 @@
     <Nullable>enable</Nullable>
   </PropertyGroup>
 
-  <ItemGroup>
-    <PackageReference Include="Microsoft.Extensions.Logging.Abstractions" Version="10.0.6" />
-  </ItemGroup>
-
 </Project>
